requirements.txt
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
-e git+ssh://git@github.com/singlestore-labs/elixirconf-demo.git@f0e07d65bf6c3ed9aed796044247606ad7296375#egg=elixirconf
idna==3.10
packaging==25.0
parsimonious==0.10.0
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.4
setuptools==80.9.0
singlestoredb==1.15.1
sqlparams==6.2.0
urllib3==2.5.0
wheel==0.45.1
